third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFECompositeElement-dom-k4-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'k4' attribute of the SVGFECompositeElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS overComposite1.getAttribute('k4') is "2.9"
PASS overComposite1.getAttribute('k4') is ".3"
PASS successfullyParsed is true

TEST COMPLETE
